[FEATURE][WIP] Support writing 'bundles' config as part of a bundle
When a bundle is created, sections can be added that will be written as
sap.ui.loader.config({bundlesUI5:{...}}) information. Each section of
type 'bundleInfo' represents one single bundle. The section name is the
bundle name, the modules in the section are the modules of the bundle.

It is currently not yet possible to reuse existing bundle definitions to
define the content of such a section.